move/split enumerate tags patches into diagnostic
