+++ SolastaCommunityExpansion/Patches/Diagnostic/GuiSpellDefinitionPatcher.cs
@@ -2,12 +2,8 @@
 using HarmonyLib;
 using SolastaCommunityExpansion.Models;
 
-namespace SolastaCommunityExpansion.Patches.GameUi
+namespace SolastaCommunityExpansion.Patches.Diagnostic
 {
-    //
-    // custom resources enablement patch
-    //
-
     [HarmonyPatch(typeof(GuiSpellDefinition), "EnumerateTags")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class GuiSpellDefinition_EnumerateTags
@@ -23,24 +19,6 @@ namespace SolastaCommunityExpansion.Patches.GameUi
             else if (DiagnosticsContext.IsCeDefinition(__instance.BaseDefinition))
             {
                 // Not all CE spells are registered in SpellsContext
-                TagsDefinitions.AddTagAsNeeded(__instance.TagsMap,
-                    "CommunityExpansion", TagsDefinitions.Criticity.Normal, true);
-            }
-        }
-    }
-
-    //
-    // custom resources enablement patch
-    //
-
-    [HarmonyPatch(typeof(GuiPowerDefinition), "EnumerateTags")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class GuiPowerDefinition_EnumerateTags
-    {
-        public static void Postfix(GuiPowerDefinition __instance)
-        {
-            if (DiagnosticsContext.IsCeDefinition(__instance.BaseDefinition))
-            {
                 TagsDefinitions.AddTagAsNeeded(__instance.TagsMap,
                     "CommunityExpansion", TagsDefinitions.Criticity.Normal, true);
             }
